+++ driver/web/adapter.go
@@ -36,6 +36,8 @@ type Adapter struct {
 	laundryapiHandler  rest.LaundryApisHandler
 	buildingapiHandler rest.BuildingAPIHandler
 	contactapiHandler  rest.ContactInfoApisHandler
+	coursesapiHandler  rest.CourseApisHandler
+	termsapiHandler    rest.TermSessionAPIHandler
 	tokenAuth          *TokenAuth
 	app                *core.Application
 }
@@ -96,6 +98,11 @@ func (we Adapter) Start() {
 
 	mainRouter.HandleFunc("/person/contactinfo", we.tokenAuthWrapFunc(we.contactapiHandler.GetContactInfo)).Methods("GET")
 
+	mainRouter.HandleFunc("/courses/giescourses", we.tokenAuthWrapFunc(we.coursesapiHandler.GetGiesCourses)).Methods("GET")
+	mainRouter.HandleFunc("/courses/studentcourses", we.tokenAuthWrapFunc(we.coursesapiHandler.GetStudentcourses)).Methods("GET")
+
+	mainRouter.HandleFunc("/termsessions/listcurrent", we.tokenAuthWrapFunc(we.termsapiHandler.GetTermSessions)).Methods("GET")
+
 	log.Fatal(http.ListenAndServe(":"+we.port, router))
 }
 
@@ -139,10 +146,13 @@ func NewWebAdapter(host string, port string, app *core.Application, tokenAuth *T
 	laundryapiHandler := rest.NewLaundryApisHandler(app)
 	buildingapiHandler := rest.NewBuildingAPIHandler(app)
 	contactapiHandler := rest.NewContactInfoApisHandler(app)
+	coursesapiHandler := rest.NewCourseApisHandler(app)
+	termsapiHandler := rest.NewTermSessionAPIHandler(app)
+
 	return Adapter{host: host, port: port,
 		apisHandler: apisHandler, adminApisHandler: adminApisHandler, app: app, laundryapiHandler: laundryapiHandler,
 		buildingapiHandler: buildingapiHandler, tokenAuth: tokenAuth,
-		contactapiHandler: contactapiHandler}
+		contactapiHandler: contactapiHandler, coursesapiHandler: coursesapiHandler, termsapiHandler: termsapiHandler}
 }
 
 //AppListener implements core.ApplicationListener interface